Dockerfile for building and creating poppler tarball (#5)
As mentioned in https://github.com/hackclub/heroku-buildpack-poppler/pull/3,
it turns out the published tarball from the poppler site is just the
source code. See that PR description for why we are packaging prebuilt
binaries instead of rebuilding from scratch.

This PR introduces a Dockerfile for creating a reproducible build that
is committed to the git repo. We are using tar metadata recommended by
https://reproducible-builds.org/docs/archives/ for consistent checksums.

Dependencies are based on https://www.linuxfromscratch.org/blfs/view/cvs/general/poppler.htm
and error/warning messages while running cmake. Due to an issue with
heroku slug size, we are using as few dependencies as possible and only
including the `pdftoppm` utility. See
https://github.com/hackclub/heroku-buildpack-poppler/pull/4 for more
reasoning about the files included here.